Add IsUnique flag to all characters
Until now, only characters at spawn points had an IsUnique flag.

With this change, the flag is moved from the SpawnPointCharacters
table to the Characters table.

A character is marked as IsUnique if they (i) ARE a prefab and only
have a single spawn point or (ii) are NOT a prefab and only a single
instance with the name exists.