fix: resolve primary key conflicts
Various ID columns that we extract 1:1 from the game are not
actually guaranteed to unique (they are manually assigned by the
dev, which occasionally causes conflicts after they copy-paste
to create a new item/spell/skill).

To avoid these "ID" conflicts from causing issues in our own
processing code, we're now using the corresponding ResourceNames
instead, which ARE guaranteed to be unique.